% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_du.R
\name{calc_du}
\alias{calc_du}
\title{Definition of the discrimination unit (DU) for setting levels}
\usage{
calc_du(fendata)
}
\arguments{
\item{fendata}{a dataframe with phenotypic quantitative data. It
must contain the columns Collection, Accession, Year, Obs, Value,
colvaryear, colvar.}
}
\value{
dataframe with the mean accession values and the range of values.
It contains the columns colvar, Collection, Accession, Mean, Rg, Rg_fil.
}
\description{
The function calculates the minimum value that allows distinction of
levels within a phenotypic characteristic using the methodology
described in Royo et al (2017). The function provides the following
outcomes in the console:
- Nacc : The number of accessions with characterization data
- Nacc_ref : The number of accessions used to calculate the discriminant
unit (UD). Accessions with highly variable observations are not considered.
- vm_g: lowest mean accesion value observed in the accessions in Nacc.
- VM_g: highest mean accession value observed in the accesions in Nacc.
- DU: Discriminant Unit, i.e. the difference between levels for the
phenotypic characteristic.
- P50: value of the 50 percentile of the mean accession values in Nacc.
- Central class: the limits of the central class defined by P50+-DU/2.
}
\details{
Additionally, the function provides a dataframe with the mean accession
values, and the observed ranges of values for them. For each accesion the
dataframe provides:
- Mean: mean of the observed values
- Rg: the difference between the maximum and minimum observed value
- Rg_fil: the difference between the maximum and minimum observed value
for all the accessions used to calculate the DU. The accessions with high
internal variability have a N/A instead.

NOTE: the current version of the function groups accessions using colvar
identifier. That is, accessions with data from two or more collecctions
are considered different. In order to compute a single accession value in
those situations, accessions should be encoded as belonging to the same
collection. Future versions of the function will allow to modify the
accession clustering method (by accession, by accession and collection).
}
\examples{
# Remove outliers using a thr_y=1.5 and thr_g =1.8
library(tidyverse)
Claw.clean <- clean_outliers(U35_PE, 1.5, 1.8)
# Define the central class and DU for the accessions
Claw.class <- calc_du(Claw.clean)
}
\references{
Royo B (Coord), 2017. Harmonized methodology for the pomological characterization
of apple (Malus x domestica Borkh.). Bilingual edition Spanish/English.
Instituto Nacional de Investigacion y Tecnologia Agraria y Alimentaria (INIA).
Madrid, Spain. 122 pp. ISBN 978-84-7498-577-1
}
\author{
Carlos Miranda, \email{carlos.miranda@unavarra.es}
}
