% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_outliers.R
\name{clean_outliers}
\alias{clean_outliers}
\title{Detection and removal of outliers within accessions}
\usage{
clean_outliers(fendata, thr_y = 1.5, thr_g = 1.5)
}
\arguments{
\item{fendata}{a dataframe with phenotypic quantitative data. It
must contain the columns Collection, Accession, Year, Obs, Value.}

\item{thr_y}{threshold for outliers within yearly observations. By
default = 1.5}

\item{thr_g}{threshold for globla outliers within accessions. By
default = 1.5}
}
\value{
dataframe in which the observations considered as outliers have
been removed, leaving a NA instead. It contains the columns Collection,
Acession, Year, Obs, Value, colvaryear, colvar
}
\description{
The function removes phenotypic observations that could be considered
as outliers from a characterization dataset based on the methodology
described in Royo et al (2017). Briefly, the characterization data for
each accession is analyzed in order to remove the observations that
deviate markedly from the others. This is a two-step process in which
first the outliers within each year of observations are eliminated,
and then at the global level (all years of observation pooled). By
default, observations differing 1.5 standard deviations from the
yearly or pooled mean are considered outliers, but the thresholds
can be adjusted by the user.
}
\examples{

# Remove outliers using a thr_y=1.5 and thr_g =1.8
library(tidyverse)
Claw.clean <- clean_outliers(U35_PE, 1.5, 1.8)
}
\references{
Royo B (Coord), 2017. Harmonized methodology for the pomological characterization
of apple (Malus x domestica Borkh.). Bilingual edition Spanish/English.
Instituto Nacional de Investigacion y Tecnologia Agraria y Alimentaria (INIA).
Madrid, Spain. 122 pp. ISBN 978-84-7498-577-1
}
\author{
Carlos Miranda, \email{carlos.miranda@unavarra.es}
}
