\name{U35_PE}
\alias{U35_PE}
\docType{data}
\title{
Characterization data for length of claw of petal (U35)
}
\description{
Characterization dummy data for length of claw of petal in pear flowers, corresponding to descriptor U35 in TG/15/3 (UPOV, 2000). The dataset contains lengths for 52 accessions, evaluated for up to three years. Simulated data included to test the functions in phenoclass
}
\usage{data("U35_PE")}
\format{
  A data frame with 1000 observations on the following 5 variables.
  \describe{
    \item{\code{Collection}}{character vector, name of the collection}
    \item{\code{Accession}}{character vector, code of the accession}
    \item{\code{Year}}{numeric vector, year of observation}
    \item{\code{Obs}}{numeric vector, number of the observation}
    \item{\code{Value}}{numeric vector, length of claw, in mm}
  }
}
\details{
Dummy data created just to provide a working example of data formatting.
}
\source{
UPOV, 2000. TG/15/3. Guidelines for the conduct of tests for distinctness, uniformity and stability for pear (Pyrus communis L.). International Union for the Protection of New Varieties of Plants. Geneva, Switzerland.
}


